requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
decorator==5.1.1
Django==5.1.3
django-celery-beat==2.7.0
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
executing==2.1.0
idna==3.10
inflection==0.5.1
ipython==8.29.0
jedi==0.19.2
kombu==5.4.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.10.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
stripe==11.3.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
